+++ src/pages/account/trips/trip-details.tsx
@@ -1,4 +1,4 @@
-import TripDetails from '@features/trip/edit/TripDetails';
+import TripDetails from '@features/trip/edit/components/TripDetails';
 
 export default function TripDetailsPage() {
   return <TripDetails />;
